chore: extracted/renamed AirflowMonitor->MonitorComponentManager
