+++ modules/dbnd-airflow-monitor/src/airflow_monitor/multiserver/multiserver.py
@@ -2,23 +2,19 @@ import logging
 
 from datetime import timedelta
 from time import sleep
-from typing import Any, Callable, Dict, List, Type, Union
+from typing import Dict, List, Type, Union
 from uuid import UUID
 
-import airflow_monitor
-
-from airflow_monitor.common import MonitorState, capture_monitor_exception
 from airflow_monitor.common.config_data import AirflowServerConfig
 from airflow_monitor.config import AirflowMonitorConfig
-from airflow_monitor.data_fetcher import get_data_fetcher
-from airflow_monitor.fixer.runtime_fixer import start_runtime_fixer
+from airflow_monitor.multiserver.monitor_component_manager import (
+    MonitorComponentManager,
+)
 from airflow_monitor.multiserver.runners import (
     RUNNER_FACTORY,
     MultiProcessRunner,
     SequentialRunner,
 )
-from airflow_monitor.multiserver.runners.base_runner import BaseRunner
-from airflow_monitor.syncer.runtime_syncer import start_runtime_syncer
 from airflow_monitor.tracking_service import (
     get_servers_configuration_service,
     get_tracking_service,
@@ -41,130 +37,11 @@ from urllib3.exceptions import HTTPError
 
 logger = logging.getLogger(__name__)
 
-KNOWN_COMPONENTS = {
-    "state_sync": start_runtime_syncer,
-    # "xcom_sync": Component,
-    # "dag_sync": Component,
-    "fixer": start_runtime_fixer,
-}
-
-
-class AirflowMonitor(object):
-    runner_factory: Callable[[Any], BaseRunner]
-
-    def __init__(
-        self,
-        server_config: AirflowServerConfig,
-        runner_factory: Callable[..., BaseRunner],
-        tracking_service: DbndAirflowTrackingService,
-    ):
-        self.server_config = server_config
-        self.active_components = {}  # type: Dict[str, BaseRunner]
-        self.tracking_service = tracking_service
-
-        self.runner_factory = runner_factory
-        self._is_stopping = False
-
-    def is_enabled(self, component_name):
-        if self._is_stopping:
-            return False
-        return getattr(self.server_config, f"{component_name}_enabled", False)
-
-    def is_active(self, component_name):
-        return component_name in self.active_components
-
-    def _update_component_state(self):
-        self._clean_dead_components()
-
-        airflow_alive = None
-        for name, runnable in KNOWN_COMPONENTS.items():
-            if self.is_enabled(name) and not self.is_active(name):
-                if airflow_alive is None:
-                    # check it only once per iteration and only if need to start anything
-                    airflow_alive = self.is_airflow_server_alive()
-                    if not airflow_alive:
-                        logger.warning(
-                            "Airflow Server is not responsive, will skip starting new syncers"
-                        )
-                if airflow_alive:
-                    self.active_components[name] = self.runner_factory(
-                        target=runnable,
-                        tracking_service=self.tracking_service,
-                        tracking_source_uid=self.server_config.tracking_source_uid,
-                    )
-                    self.active_components[name].start()
-            elif not self.is_enabled(name) and self.is_active(name):
-                component = self.active_components.pop(name)
-                component.stop()
-
-        return airflow_alive
-
-    def _clean_dead_components(self):
-        for name, component in list(self.active_components.items()):
-            if not component.is_alive():
-                component.stop()
-                self.active_components.pop(name)
-
-    @capture_monitor_exception("stopping monitor")
-    def stop(self):
-        self._is_stopping = True
-        self._update_component_state()
-
-    @capture_monitor_exception("starting monitor")
-    def start(self):
-        self.tracking_service.update_monitor_state(
-            MonitorState(
-                monitor_start_time=utcnow(),
-                monitor_status="Scheduled",
-                airflow_monitor_version=airflow_monitor.__version__,
-                monitor_error_message=None,
-            ),
-        )
-        alive = self._update_component_state()
-
-        if alive:
-            plugin_metadata = get_data_fetcher(self.server_config).get_plugin_metadata()
-        else:
-            plugin_metadata = None
-
-        self.tracking_service.update_monitor_state(
-            MonitorState(
-                monitor_start_time=utcnow(),
-                monitor_status="Running",
-                airflow_monitor_version=airflow_monitor.__version__,
-                airflow_version=plugin_metadata.airflow_version
-                if plugin_metadata
-                else None,
-                airflow_export_version=plugin_metadata.plugin_version
-                if plugin_metadata
-                else None,
-            ),
-        )
-
-    @capture_monitor_exception("updating monitor config")
-    def update_config(self, server_config: AirflowServerConfig):
-        self.server_config = server_config
-        self._update_component_state()
-
-    @capture_monitor_exception("checking monitor alive")
-    def is_airflow_server_alive(self):
-        return get_data_fetcher(self.server_config).is_alive()
-
-    @capture_monitor_exception("heartbeat monitor")
-    def heartbeat(self):
-        for component in self.active_components.values():
-            component.heartbeat()
-
-        self._clean_dead_components()
-
-    def __str__(self):
-        return f"AirflowMonitor({self.server_config.name}|{self.server_config.tracking_source_uid})"
-
 
 class MultiServerMonitor(object):
     runner_factory: Union[Type[SequentialRunner], Type[MultiProcessRunner], None]
     monitor_config: AirflowMonitorConfig
-    active_monitors: Dict[UUID, AirflowMonitor]
+    active_monitors: Dict[UUID, MonitorComponentManager]
     tracking_service: DbndAirflowTrackingService
 
     def __init__(
@@ -275,7 +152,7 @@ class MultiServerMonitor(object):
                 monitor.update_config(server_config)
             else:
                 logger.info(f"Starting new monitor for {server_id}")
-                monitor = AirflowMonitor(
+                monitor = MonitorComponentManager(
                     server_config,
                     self.runner_factory,
                     tracking_service=get_tracking_service(server_id),
@@ -287,7 +164,7 @@ class MultiServerMonitor(object):
         servers_ids = {s.tracking_source_uid for s in servers_configs}
         for server_id, monitor in list(
             self.active_monitors.items()
-        ):  # type: (UUID, AirflowMonitor)
+        ):  # type: (UUID, MonitorComponentManager)
             if server_id not in servers_ids:
                 logger.info(f"Server {server_id} not longer enabled, going to stop it")
                 monitor.stop()
